6_Identity_AuthN_n_AutZ_n_Users_n_1toMany_n_Pagination_n_403_Error_Handler

+++ WebApplication1/Services/PlaceService/PlaceServiceImpl.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Microsoft.EntityFrameworkCore;
 using WebApplication1.Data;
 using WebApplication1.DTO;
@@ -13,11 +14,14 @@ namespace WebApplication1.Services
         private readonly ICoordinatesService _coordinatesService;
         private readonly IImageService _imageService;
 
-        public PlaceServiceImpl(DataContext context, ICoordinatesService coordinatesService, IImageService imageService)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+        public PlaceServiceImpl(DataContext context, ICoordinatesService coordinatesService, IImageService imageService, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
             _coordinatesService = coordinatesService;
             _imageService = imageService;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public async Task<List<Place>> GetPlaces()
@@ -25,8 +29,41 @@ namespace WebApplication1.Services
             return await _context.Places.ToListAsync();
         }
 
+        public async Task<List<Place>> GetPagedPlaces(int pageNumber, int pageSize)
+        {
+            return await _context.Places
+                .Skip((pageNumber - 1) * pageSize)
+                .Take(pageSize)
+                .ToListAsync();
+        }
+
+        public async Task<int> GetPlacesCount()
+        {
+            return await _context.Places.CountAsync();
+        }
+
+        public async Task<int> GetPlacesCountByUserId(string userId)
+        {
+            return await _context.Places.CountAsync(p => p.Creator == userId);
+        }
+
+        public async Task<List<Place>> GetPagedPlacesByUserId(string userId, int pageNumber, int pageSize)
+        {
+            return await _context.Places
+                .Where(p => p.Creator == userId)
+                .Skip((pageNumber - 1) * pageSize)
+                .Take(pageSize)
+                .ToListAsync();
+        }
+
         public async Task AddPlace(Place place, IFormFile? image)
         {
+            if (_httpContextAccessor.HttpContext != null)
+            {
+                var userId = _httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+                place.Creator = userId; // assign logged in user as creator
+            }
+
             if (image != null)
             {
                 string uniqueFileName = Guid.NewGuid().ToString();
@@ -54,39 +91,50 @@ namespace WebApplication1.Services
             var place = await _context.Places.FindAsync(id);
             if (place != null)
             {
-                place.Title = updatedPlace.Title ?? place.Title;
-                place.Description = updatedPlace.Description ?? place.Description;
-                place.PlaceLocation = updatedPlace.PlaceLocation ?? place.PlaceLocation;
-                place.Address = updatedPlace.Address ?? place.Address;
-                place.Creator = updatedPlace.Creator ?? place.Creator;
-                place.ImageUrl = updatedPlace.ImageUrl ?? place.ImageUrl;
-
-                if (image != null)
+                if (_httpContextAccessor.HttpContext != null)
                 {
-                    // unlink existing image
-                    if (!string.IsNullOrEmpty(place.ImageUrl))
+                    var userId = _httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+                    // check if logged in user owns the place
+                    if (place.Creator == userId)
                     {
-                        var oldImageName = new Uri(place.ImageUrl).Segments.Last();
-                        await _imageService.DeleteImageAsync(oldImageName);
+                        place.Title = updatedPlace.Title ?? place.Title;
+                        place.Description = updatedPlace.Description ?? place.Description;
+                        place.PlaceLocation = updatedPlace.PlaceLocation ?? place.PlaceLocation;
+                        place.Address = updatedPlace.Address ?? place.Address;
+                        place.Creator = updatedPlace.Creator ?? place.Creator;
+                        place.ImageUrl = updatedPlace.ImageUrl ?? place.ImageUrl;
+                        if (image != null)
+                        {
+                            // unlink existing image
+                            if (!string.IsNullOrEmpty(place.ImageUrl))
+                            {
+                                var oldImageName = new Uri(place.ImageUrl).Segments.Last();
+                                await _imageService.DeleteImageAsync(oldImageName);
+                            }
+                            // upload new image
+                            string uniqueFileName = Guid.NewGuid().ToString();
+                            using var fileStream = image.OpenReadStream();
+                            var imageUrl = await _imageService.UploadImageAsync(uniqueFileName, fileStream, image.ContentType);
+                            place.ImageUrl = imageUrl;
+                        }
+                        if (!string.IsNullOrEmpty(updatedPlace.Address))
+                        {
+                            var geolocationResponse = await _coordinatesService.GetCoordinatesForAddressAsync(updatedPlace.Address);
+                            if (geolocationResponse != null)
+                            {
+                                var location = geolocationResponse;
+                                place.PlaceLocation = new PlaceLocation(location.Lat, location.Lng);
+                            }
+                        }
+                        _context.Places.Update(place);
+                        await _context.SaveChangesAsync();
                     }
-                    // upload new image
-                    string uniqueFileName = Guid.NewGuid().ToString();
-                    using var fileStream = image.OpenReadStream();
-                    var imageUrl = await _imageService.UploadImageAsync(uniqueFileName, fileStream, image.ContentType);
-                    place.ImageUrl = imageUrl;
+                    else
+                    {
+                        throw new UnauthorizedAccessException();
+                    }
                 }
 
-                if (!string.IsNullOrEmpty(updatedPlace.Address))
-                {
-                    var geolocationResponse = await _coordinatesService.GetCoordinatesForAddressAsync(updatedPlace.Address);
-                    if (geolocationResponse != null)
-                    {
-                        var location = geolocationResponse;
-                        place.PlaceLocation = new PlaceLocation(location.Lat, location.Lng);
-                    }
-                }
-                _context.Places.Update(place);
-                await _context.SaveChangesAsync();
             }
         }
 
@@ -95,15 +143,27 @@ namespace WebApplication1.Services
             var place = await _context.Places.FindAsync(id);
             if (place != null)
             {
-                // unlink existing image
-                if (!string.IsNullOrEmpty(place.ImageUrl))
+                if (_httpContextAccessor.HttpContext != null)
                 {
-                    var oldImageName = new Uri(place.ImageUrl).Segments.Last();
-                    await _imageService.DeleteImageAsync(oldImageName);
+                    var userId = _httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+                    // check if logged in user owns the place
+                    if (place.Creator == userId)
+                    {
+                        // unlink existing image
+                        if (!string.IsNullOrEmpty(place.ImageUrl))
+                        {
+                            var oldImageName = new Uri(place.ImageUrl).Segments.Last();
+                            await _imageService.DeleteImageAsync(oldImageName);
+                        }
+                        _context.Places.Remove(place);
+                        await _context.SaveChangesAsync();
+                    }
+                    else
+                    {
+                        throw new UnauthorizedAccessException();
+                    }
                 }
 
-                _context.Places.Remove(place);
-                await _context.SaveChangesAsync();
             }
         }
     }
